+++ include/cub3d.h
@@ -20,3 +20,16 @@
 /* ----- RULES ----- */
 #define GREETING true
 #define VERBOSE true
+
+/* ---- FUNCTIONS ----- */
+// lets mirror the file system in here. the less ---, the deeper the file
+
+// ----- util
+// garbage collector
+t_list		*gc_create(void);
+int			gc_append_element(t_list *gc, void *content);
+void		gc_cleanup_and_reinit(t_list **gc);
+void		*gc_malloc(size_t size, t_list **gc, void *ptr1);
+// printing
+void		logger(char type, char *message);
+void		verbose_logger(char type, char *message);