+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### [0.1.0] - 21.05.2024 - Freddy
+- Added duplicate protected garbage collector with gc_malloc
+- added logger and verbose logger
+- makefile object folder subdirectories fix
+- makefile not properly cleaning submodules fix
+
 ## [0.0.0] - 21.05.2024 - Freddy
 - initial commit
 - Basic project setup